requirements.txt
meilisearch>=0.9.0
python-slugify>=4.0.0
